requirements.txt
llama-index >= 0.5.0
slack-bolt >= 1.6.0
python-slugify >= 8.0.0
GitPython > 3.1.0
faiss-cpu >= 1.6.3
langchain >= 0.0.147
chromadb >= 0.3.21


